Enable PASTE for pw field on 'unlock wallet' screen
Password field on 'unlock wallet' screen is the only password field in
application that has PASTE option disabled. So, users most likely will
expect that they can paste password in this field too.
https://github.com/KomodoPlatform/atomicDEX/issues/132#issuecomment-579244059

There is also idea to make single PASTE option enabled only if
password field is empty (and hide CUT, COPY and SELECT ALL options).
But in order to implement it, we need to use 'toolbarOptions'
property, because changing 'enableInteractiveSelection' flag in
setState does not have any effect. And 'toolbarOptions' only works
properly in flutter v.1.10.13+
https://github.com/flutter/flutter/issues/45534
https://github.com/flutter/flutter/commit/b730a993c1e6dd2774b270fb2e0253a26e69cf39
So, we can come back to this later, after flutter version will
be updated (current is v1.9.1+hotfix.6)

#644