Wait indefinitely for authenticateWithBiometrics
I've seen the app "interrupted" on Android when I've 1) triggered the fingerprint prompt; 2) went to home screen; 3) later went back to the app. Hypothesis is that the biometrics plugin can crash when we invoke it the second time. Hence we now prefer to wait for the biometrics to finish before invoking it again, even if it takes a long time.