Exit when backgrounded at the end of the audio life-line
There's a chance that our application is being suspended when it stops playing the sounds.

Our previous encounters with this (https://gitlab.com/artemciy/supernet/issues/4) seems to indicate that suspended app termination is a gradual process, open to funny side effects.

In this patch we `exit` the backgrounded application right away in case we aren't going to enjoy the saving grace of the sound for long.

#658