Account for late lifecycle events
Our application goes into background when the user leaves it and goes into the faceid biometrics system dialogue. When user returns back from faceid, under iOS 12 we get notified of coming back to foreground, then later are notified of the successfull biometrics check, after which we switch `showLock` off. On iOS the order of these events is usually reversed, so we've been unlocking the screen, because the biometrics passes, and then locking it back again, because the application was in background.

The patch prevents this unfortunate ordering of events from triggering the re-lock by ignoring background transitions which happen immediately after the biometrics.

#672