Improve HTTP handling in getBalance
HTTP has several error modes. For example, an application might use status code 200 to return an error, which is correct behavior because application error is a different *layer* and should in general be handled independent of the transport layer erros. But an application might also return an error with status 500, which is also valid. Except status 500 might be coming from a web server or other layers and not from an application, hence it should be handled with more care than status 200.

With this commit the error handling in getBalance looks sufficiently thorough to me. Ping me if you see a space for improvements.

We should probably see how it works in the wild before refactoring the rest of the HTTP handling methods.

This commit also switches to the specific return type (Balance) and to throwing the errors, greatly simplifying the code.
This can affect getAllBalances in mm_service. We can further improve getAllBalances if it proves to have hitches.

#665