Remove old swap model code
This should also fix some swap timeouts, the old code was timing out swaps indiscriminately.

We shouldn't need to time out the swaps from UI, MM will do that for us.
But if, for some reason, we would still want to do UI timeouts in the future, then we should take swap progress and swap speed predictions into account.

Also in this commit a number of shortcut refactorigs (as Flutter's enforcing a small line width).

#638