Fix running tests with "flutter test"
Most ways of running the unit tests in flutter set the current
directory as the root of the project, but "flutter test" sets it to
the "test/" folder.

This changes the fixture function to assume the current directory as
the "test/" folder, effectively making "flutter test" work while
breaking other ways of testing, such as using the IDE and
"flutter test" passing the name of the file containing the tests.