Revamp log rotation
Simplified the log structure a bit: now every day has a separate log file. This has reliability benefits, since we never replace the log "in place".

It should now be possible to run the application indefinitely without it ever exhausing the log space (#653), as it will create new files for new logging days and will also remove the logs for the old previous days.

Log sharing now picks a 33 MiB portion of the log and compresses it with gzip. That way it should always fit into a Discord attachment.

We're also fetching the BUILD_TIME for #689.