Refactor MM update
This time we're using the BUILD_TIME in order to skip some of the assets/mm2 checks.

I don't think we want updateMmBinary to be catching all exceptions: if we can't have the copy of the fresh MM version then something is likely wrong and we shouldn't be running that instance until the problem is resolved.

#689