Do not copy coins to DB
With this patch the JSON files, coins_init_mm2.json and coins_config.json, are directly used to load the fresh information about the coins. That way we're storing most coin information in one place (instead of two or three), which makes updating it easier. That is, settings like txfee and the number of confirmations can now be updating by simply overwriting coins_init_mm2.json with a fresh copy from github.com/jl777/coins.

#689