Start gossiping metrics
The system metrics (application network usage (#349), RAM, priority, etc) will give us an extra signal about the Maker health (#648).

For example, the chances of a successful swap might be lower if the Maker didn't get the "foreground service" priority on Android, or if it's leaking memory and grew in size.

This will also allow us to detect file and memory leaks and to detect the changes in resource utilization between the different versions.