Fix disable_coin tests
In order to proceed with base91 testing I need the flutter tests fixed. First failing test is disable_coin. To fix it this commit provides the tested application with a documents directory, allowing it to save logs. Also refactors disable_coin to use the improved HTTP error handling, streamlines the asynchronous code path and removes the unnecessary error conditions.

#727