Make add-coins button collapsed change clearer
Change the animation of the appearance and disappearance of the collapsed add-coins icon button to make it flash-like.

Only change to collapsed state once the un-collapsed button is entirely off-screen so that the user is not overwhelmed with too many simultaneous animations and does not notice the add-coins button appear in the app bar.