Change PrimaryButton sizing/styling
Remove the padding from the PrimaryButton because the vertical padding parameter was used as a simple fix to have the button have a bigger height than the default.

However, this interfered with the widget’s responsiveness by cutting off multi-line text and not aligning the button content in the center. This was changed rather to set the min height property because it achieves the desired effect with minimal effect on the widget’s default layout.