Refactor authentication for local authentication
Refactor authentication repository in preparation for implementing system-controlled biometrics authentication.

This changes the original strategy, where we handle various sign-in methods. The new approach is to pass this responsibility to the system. Android and iOS have pretty cool features where we can store secure data, which requires biometric authentication to read.