Create App logo widget
Create app logo widget which shows the logo in multiple styles. See constructors for more info.

Will take the size constraints which its parent imposes on it. If the constraints are unbounded, it will revert to a height of 40 for full logo and 24 for icon logo.

In general, it’s a good practice avoid setting specific sizes of reusable widgets unless it doesn’t make sense. Let the parent determine the constraints. This allows our reusable widgets to be more “reusable”.

See Flutter docs on [LimitedBox] for more info.