New app root widget and nav system
Refactor app root widget to new widget, integrating with new navigation and bloc architecture. The aim is also to delegate as much as possible to the system with the upcoming refactoring. E.g. Theming, localisation selection, and authentication.

Note that the code is copied and modified from `main.dart`. The commented-out code has been left in the commit to ensure that all the legacy functionality is accounted for by refactoring into a new stack or removing it because it is no longer needed.