Create A-Dex Account model
Create an Account ID and Account model following SOLID principles to represent the identity of an account.

Each account type and its ID have a separate data structure, which deals with account IDs in the UI. With this approach, the ID can be represented as a model, and the == operated can be used to check for equality instead of complex logic with strings.