Implement middleware overlay for biometrics
Create a middleware widget which sits between the Material App’s router and its route builder. Used for global overlays or to listen globally for bloc state changes. See widget comments.

Use middleware widget to implement a biometrics status widget which informs the user if their device does not support biometrics or if they have not set it up.