Implement app bar and legacy routes with beamer
Add the legacy routes to the new navigation system.

This only adds them as top-level routes to the widget for each page. Once on the page, the page is responsible for it’s own sub-routes with Nav 1.0 pop and pushing. In the long-term, we should have all nested routes as a location in the new nav system so that we can easily navigate between parts of the app and create deep-links. This also allows us to store the exact navigator state the app is in.