Change sidebar “logout” to new bloc
Update the sidebar “logout” button to use the new bloc instead of the old auth system.

Remove the `context` parameter from the widget. Avoid passing context as a parameter to a widget since the widget has its own context. If we’re trying to access widgets in other parts of the widget tree, consider a different approach instead of passing context.