Create a re-usable eager loading button
Create a re-usable Floating Action Button (FAB) which when tapped while not ready, will display a circular progress loader. If the FAB becomes ready after being tapped, the provided `onTap` callback will be triggered automatically.

This is useful to give the user the idea that the loading isn’t happening unless they tap the button before the loading is complete.

This is created so that it can be used for the Add coin FAB.