Bug - Fix only first 10 transactions showing
- Fix bug where the coins page would only load the first 10 transactions and not load more transactions when scrolled to the bottom of the list.

There is also other refactoring for the coin details screen, but the root cause of the issue was improper use of layouts scrollable areas.

NB! A significant portion of the code in this commit’s diff is unchanged except for being moved as part of the refactoring. Dig deeper into the commit history for more info on the origin of this code. Using helper widget methods instead of separate widgets is not a good practice.