Bug - Make coin update run on every start
Previously, the updater would run only if the update was not already downloaded. I.e. The updater would run only on the initial launch after installing the app.

Now, the update will run each time the app launches. As was before, the downloaded update only takes effect from the next launch. This reduces launch time and prevents inconsistencies if the data is fetched after the update is run and other parts of the app reference it before the update runs.

Also throw in some extra error handling just in case.

Diamonds are forever… just like a dev that plans no further commits.